refactor: moved components
